feat(Data/UInt, Data/Fin/Basic) instances
Add new typeclass instances for Fin and UInt8. This incorporates gebner's suggestions to add boolean methods for checking under/overflow, and adding separate instances of e.g. `Fin UInt8.size`. The plan was to make a macro that would extend the UInt stuff to all the UInt sizes, but I couldn't figure out how to deal with namespaces and identifiers properly in a macro, so I'll have to get some outside guidance and improve the macro docs.